+++ backend/src/helpers/generateCookie.js
@@ -5,6 +5,6 @@ export default function generateCookie(res, user) {
         httpOnly: true,
         secure: false,
         sameSite: 'strict',
-        maxAge: user.isRemembered ? 30 * 24 * 60 * 60 * 1000 : 24 * 60 * 60 * 1000
+        maxAge: user.isRemembered ? 30 * 24 * 60 * 60 * 1000 : 24 * 60 * 60 * 1000,
     });
 }