+++ backend/src/index.js
@@ -5,14 +5,8 @@ import express from 'express';
 import cors from 'cors';
 import cookieParser from 'cookie-parser';
 import mongoose from 'mongoose';
-import authRoutes from './routes/auth.routes.js';
-import userRoutes from './routes/user.routes.js';
-import accessorieRoutes from './routes/accessorie.routes.js';
-import assetRoutes from './routes/asset.routes.js';
-import licenseRoutes from './routes/license.routes.js';
-import locationRoutes from './routes/location.routes.js';
-import companyRoutes from './routes/company.routes.js';
-import departmentRoutes from './routes/department.routes.js';
+import Startup from './startup/index.js';
+import Routes from './routes/index.js';
 import authMiddleware from './middleware/auth.middleware.js';
 
 const app = express();
@@ -29,20 +23,23 @@ app.use(cookieParser());
 
 // routes
 app.use('/api/status', (req, res) => res.send('OK'));
-app.use('/api/auth', authRoutes);
-app.use('/api/users', authMiddleware, userRoutes);
-app.use('/api/accessories', authMiddleware, accessorieRoutes);
-app.use('/api/assets', authMiddleware, assetRoutes);
-app.use('/api/licenses', authMiddleware, licenseRoutes);
-app.use('/api/location', authMiddleware, locationRoutes);
-app.use('/api/company', authMiddleware, companyRoutes);
-app.use('/api/department', authMiddleware, departmentRoutes);
+app.use('/api/auth', Routes.AuthRoutes);
+app.use('/api/accessories', authMiddleware, Routes.AccessorieRoutes);
+app.use('/api/assets', authMiddleware, Routes.AssetRoutes);
+app.use('/api/licenses', authMiddleware, Routes.LicenseRoutes);
+app.use('/api/location', authMiddleware, Routes.LocationRoutes);
+app.use('/api/company', authMiddleware, Routes.CompanyRoutes);
+app.use('/api/department', authMiddleware, Routes.DepartmentRoutes);
 
 // connections
 mongoose
     .connect(process.env.MONGO_URI)
-    .then(() => {
+    .then(async () => {
         const PORT = process.env.PORT || 3000;
         app.listen(PORT, () => console.log(`App is running on port ${PORT}`));
+
+        if ((await Startup.isInitialized()) || process.env.GENERATE_DATA) {
+            await Startup.DataGenerator(process.env.ENVIRONMENT);
+        }
     })
     .catch((err) => console.error(err));