+++ backend/src/startup/Init.js
@@ -1,0 +1,6 @@
+import User from '../models/user.models.js';
+
+export default async () => {
+    const result = await User.countDocuments({});
+    return !result;
+};