+++ backend/src/startup/dataGenerator/index.js
@@ -1,0 +1,3 @@
+import GenerateUsers from './generateUsers.js';
+
+export default async (environment) => [await GenerateUsers(environment)];