+++ backend/src/startup/index.js
@@ -1,0 +1,7 @@
+import isInitialized from './Init.js';
+import DataGenerator from './dataGenerator/index.js';
+
+export default {
+    isInitialized,
+    DataGenerator,
+};